+++ src/__init__.js
@@ -3,8 +3,8 @@
 * @description A Platonic Solid for Ideal Data
 * @author rcj1492
 * @license MIT
-* @version 0.0.7
-* @copyright 2018-2019 Collective Acuity 
+* @version 0.0.9
+* @copyright 2018-2021 Collective Acuity 
 * @email support@collectiveacuity.com
 * @url https://github.com/collectiveacuity/d12
 */